Changed theme details UI when adding theme, cancel deletes new theme when adding.
